delete chain_header from agent state
