+++ core/src/agent/mod.rs
@@ -1,6 +1,5 @@
 /// Agent is the module that handles the user’s identity and source chain for every Phenotype.
 ///
 pub mod actions;
-pub mod chain_header;
 pub mod chain_store;
 pub mod state;
