+++ core_types/src/chain_header.rs
@@ -1,5 +1,5 @@
 use cas::content::{Address, AddressableContent, Content};
-use entry::Entry;
+use entry::{test_entry, test_entry_type, Entry};
 use entry_type::EntryType;
 use error::HolochainError;
 use json::ToJson;
@@ -71,19 +71,6 @@ impl ChainHeader {
         &self.entry_type
     }
 
-    /// returns true if the entry type is a system entry
-    pub fn is_sys(&self) -> bool {
-        match self.entry_type {
-            EntryType::App(_) => true,
-            _ => false,
-        }
-    }
-
-    /// returns true if the entry type is an app entry
-    pub fn is_app(&self) -> bool {
-        !self.is_sys()
-    }
-
     /// timestamp getter
     pub fn timestamp(&self) -> &str {
         &self.timestamp
@@ -142,23 +129,25 @@ impl AddressableContent for ChainHeader {
             .expect("could not read Json as valid ChainHeader Content")
     }
 }
-/*
+
+/// returns a dummy header for use in tests
+pub fn test_chain_header() -> ChainHeader {
+    ChainHeader::new(
+        &test_entry_type(),
+        &String::new(),
+        None,
+        &test_entry().address(),
+        &String::new(),
+        None,
+    )
+}
+
 #[cfg(test)]
 pub mod tests {
     use cas::content::{Address, AddressableContent};
-    use chain_header::{header::ChainHeader, tests::test_chain, SourceChain};
-    use hash_table::{
-        entry::tests::{
-            test_entry, test_entry_a, test_entry_b, test_entry_type, test_entry_type_a,
-            test_entry_type_b,
-        },
-        sys_entry::ToEntry,
-    };
-
-    /// returns a dummy header for use in tests
-    pub fn test_chain_header() -> ChainHeader {
-        test_chain().create_next_chain_header(&test_entry_type(), &test_entry())
-    }
+    use chain_header::{test_chain_header, ChainHeader};
+    use entry::{test_entry, test_entry_b, test_entry_type, test_entry_type_a, test_entry_type_b};
+    use to_entry::ToEntry;
 
     /// returns a dummy header for use in tests
     pub fn test_chain_header_a() -> ChainHeader {
@@ -167,7 +156,14 @@ pub mod tests {
 
     /// returns a dummy header for use in tests. different from test_chain_header_a.
     pub fn test_chain_header_b() -> ChainHeader {
-        test_chain().create_next_chain_header(&test_entry_type_b(), &test_entry_b())
+        ChainHeader::new(
+            &test_entry_type_b(),
+            &String::new(),
+            None,
+            &test_entry_b().address(),
+            &String::new(),
+            None,
+        )
     }
 
     pub fn test_header_address() -> Address {
@@ -177,274 +173,218 @@ pub mod tests {
     #[test]
     /// tests for PartialEq
     fn eq() {
-        let chain_a = test_chain();
-
-        let entry_type_a = test_entry_type_a();
-        let entry_type_b = test_entry_type_b();
-
-        let entry_a = test_entry_a();
-        let entry_b = test_entry_b();
-
-        // same content + chain state is equal
-        assert_eq!(
-            chain_a.create_next_chain_header(&entry_type_a, &entry_a),
-            chain_a.create_next_chain_header(&entry_type_a, &entry_a),
-        );
+        // basic equality
+        assert_eq!(test_chain_header(), test_chain_header());
 
         // different content is different
-        assert_ne!(
-            chain_a.create_next_chain_header(&entry_type_a, &entry_a),
-            chain_a.create_next_chain_header(&entry_type_a, &entry_b),
-        );
+        assert_ne!(test_chain_header_a(), test_chain_header_b());
 
         // different type is different
         assert_ne!(
-            chain_a.create_next_chain_header(&entry_type_a, &entry_a),
-            chain_a.create_next_chain_header(&entry_type_b, &entry_a),
+            ChainHeader::new(
+                &test_entry_type_a(),
+                &String::new(),
+                None,
+                &test_entry().address(),
+                &String::new(),
+                None
+            ),
+            ChainHeader::new(
+                &test_entry_type_b(),
+                &String::new(),
+                None,
+                &test_entry().address(),
+                &String::new(),
+                None
+            ),
         );
 
-        // different state is different with same entry
-        let mut chain_b = test_chain();
-        chain_b
-            .push_entry(&entry_type_a, &entry_a)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
+        // different previous header is different
         assert_ne!(
-            chain_a.create_next_chain_header(&entry_type_a, &entry_a),
-            chain_b.create_next_chain_header(&entry_type_a, &entry_a)
+            ChainHeader::new(
+                &test_entry_type(),
+                &String::new(),
+                None,
+                &test_entry().address(),
+                &String::new(),
+                None
+            ),
+            ChainHeader::new(
+                &test_entry_type(),
+                &String::new(),
+                Some(test_chain_header().address()),
+                &test_entry().address(),
+                &String::new(),
+                None
+            ),
         );
     }
 
     #[test]
     /// tests for ChainHeader::new()
     fn new() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
+        let chain_header = test_chain_header();
 
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        assert_eq!(header.entry_address(), &entry.address());
-        assert_eq!(header.link(), None);
-        assert_ne!(header.address(), Address::new());
+        assert_eq!(chain_header.entry_address(), &test_entry().address());
+        assert_eq!(chain_header.link(), None);
+        assert_ne!(chain_header.address(), Address::new());
     }
 
     #[test]
     /// tests for header.entry_type()
     fn entry_type() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        assert_eq!(header.entry_type(), &entry_type);
+        assert_eq!(test_chain_header().entry_type(), &test_entry_type());
     }
 
     #[test]
     /// tests for header.time()
     fn time() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        assert_eq!(header.timestamp(), "");
+        assert_eq!(test_chain_header().timestamp(), "");
     }
 
     #[test]
-    /// tests for header.next()
-    fn next() {
-        let mut chain = test_chain();
-
-        let entry_type_a = test_entry_type_a();
-        let entry_type_b = test_entry_type_b();
-
-        let entry_a = test_entry_a();
-        let entry_b = test_entry_b();
-
-        // first header is genesis so next should be None
-        let chain_header_a = chain
-            .push_entry(&entry_type_a, &entry_a)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
-        assert_eq!(chain_header_a.link(), None);
-
-        // second header next should be first header hash
-        let chain_header_b = chain
-            .push_entry(&entry_type_b, &entry_b)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
-        assert_eq!(
-            chain_header_b.link(),
-            Some(chain_header_a.to_entry().1.address())
+    fn link_test() {
+        let chain_header_a = test_chain_header();
+        let chain_header_b = ChainHeader::new(
+            &test_entry_type(),
+            &String::new(),
+            Some(chain_header_a.address()),
+            &test_entry().address(),
+            &String::new(),
+            None,
         );
+        assert_eq!(None, chain_header_a.link());
+        assert_eq!(Some(chain_header_a.address()), chain_header_b.link());
     }
 
     #[test]
-    /// tests for header.entry()
-    fn entry() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        // header for an entry should contain the entry hash under entry()
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        assert_eq!(header.entry_address(), &entry.address());
+    fn entry_test() {
+        assert_eq!(test_chain_header().entry_address(), &test_entry().address());
     }
 
-    #[test]
+    // #[test]
     /// tests for header.type_next()
-    fn type_next() {
-        let mut chain = test_chain();
-
-        let entry_type_a = test_entry_type_a();
-        let entry_type_b = test_entry_type_b();
-        let entry_type_c = test_entry_type_a();
-
-        let entry_a = test_entry_a();
-        let entry_b = test_entry_b();
-        let entry_c = test_entry_b();
-
-        // first header is genesis so next should be None
-        let chain_header_a = chain
-            .push_entry(&entry_type_a, &entry_a)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
-        assert_eq!(chain_header_a.link_same_type(), None);
-
-        // second header is a different type so next should be None
-        let chain_header_b = chain
-            .push_entry(&entry_type_b, &entry_b)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
-        assert_eq!(chain_header_b.link_same_type(), None);
-
-        // third header is same type as first header so next should be first header hash
-        let chain_header_c = chain
-            .push_entry(&entry_type_c, &entry_c)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
-        assert_eq!(
-            chain_header_c.link_same_type(),
-            Some(chain_header_a.address())
-        );
-    }
+    // fn type_next() {
+    //     let mut chain = test_chain();
+    //
+    //     let entry_type_a = test_entry_type_a();
+    //     let entry_type_b = test_entry_type_b();
+    //     let entry_type_c = test_entry_type_a();
+    //
+    //     let entry_a = test_entry_a();
+    //     let entry_b = test_entry_b();
+    //     let entry_c = test_entry_b();
+    //
+    //     // first header is genesis so next should be None
+    //     let chain_header_a = chain
+    //         .push_entry(&entry_type_a, &entry_a)
+    //         .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
+    //
+    //     assert_eq!(chain_header_a.link_same_type(), None);
+    //
+    //     // second header is a different type so next should be None
+    //     let chain_header_b = chain
+    //         .push_entry(&entry_type_b, &entry_b)
+    //         .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
+    //
+    //     assert_eq!(chain_header_b.link_same_type(), None);
+    //
+    //     // third header is same type as first header so next should be first header hash
+    //     let chain_header_c = chain
+    //         .push_entry(&entry_type_c, &entry_c)
+    //         .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
+    //
+    //     assert_eq!(
+    //         chain_header_c.link_same_type(),
+    //         Some(chain_header_a.address())
+    //     );
+    // }
 
     #[test]
-    /// tests for header.signature()
+    /// tests for chain_header.signature()
     fn signature() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        assert_eq!("", header.entry_signature());
+        assert_eq!("", test_chain_header().entry_signature());
     }
 
     #[test]
     /// test header.address() against a known value
     fn known_address() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        // check a known hash
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        assert_eq!(test_header_address(), header.address());
+        assert_eq!(
+            test_chain_header_a().address(),
+            test_chain_header().address()
+        );
     }
 
     #[test]
     /// test that different entry content returns different addresses
     fn address_entry_content() {
-        let chain = test_chain();
-
-        let entry_type_a = test_entry_type_a();
-        let entry_type_b = test_entry_type_b();
-
-        let entry_a = test_entry_a();
-        let entry_b = test_entry_b();
-
-        // different entries must return different hashes
-        let header_a = chain.create_next_chain_header(&entry_type_a, &entry_a);
-
-        let header_b = chain.create_next_chain_header(&entry_type_b, &entry_b);
-
-        assert_ne!(header_a.address(), header_b.address());
-
-        let entry_type_c = test_entry_type_a();
-        let entry_c = test_entry_a();
-
-        // same entry must return same address
-        let header_c = chain.create_next_chain_header(&entry_type_c, &entry_c);
-
-        assert_eq!(header_a.address(), header_c.address());
+        assert_ne!(
+            test_chain_header_a().address(),
+            test_chain_header_b().address()
+        );
     }
 
     #[test]
     /// test that different entry types returns different addresses
     fn address_entry_type() {
-        let chain = test_chain();
-
-        let entry_type_a = test_entry_type_a();
-        let entry_type_b = test_entry_type_b();
-
-        let entry = test_entry();
-
-        let header_a = chain.create_next_chain_header(&entry_type_a, &entry);
-        let header_b = chain.create_next_chain_header(&entry_type_b, &entry);
-
-        // different types must give different addresses
-        assert_ne!(header_a.address(), header_b.address());
+        assert_ne!(
+            ChainHeader::new(
+                &test_entry_type_a(),
+                &String::new(),
+                None,
+                &test_entry().address(),
+                &String::new(),
+                None
+            ).address(),
+            ChainHeader::new(
+                &test_entry_type_b(),
+                &String::new(),
+                None,
+                &test_entry().address(),
+                &String::new(),
+                None
+            ).address(),
+        );
     }
 
     #[test]
     /// test that different chain state returns different addresses
     fn address_chain_state() {
-        // different chain, different address
-        let mut chain = test_chain();
-
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        let chain_header_control = chain.create_next_chain_header(&entry_type, &entry);
-
-        let chain_header_a = chain
-            .push_entry(&entry_type, &entry)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-        // p2 will have a different address to p1 with the same entry as the chain state is different
-        let chain_header_b = chain
-            .push_entry(&entry_type, &entry)
-            .expect("pushing a valid entry to an exlusively owned chain shouldn't fail");
-
-        assert_eq!(chain_header_control.address(), chain_header_a.address());
-        assert_ne!(chain_header_control.address(), chain_header_b.address());
+        assert_ne!(
+            test_chain_header().address(),
+            ChainHeader::new(
+                &test_entry_type_a(),
+                &String::new(),
+                Some(test_chain_header().address()),
+                &test_entry().address(),
+                &String::new(),
+                None
+            ).address(),
+        );
     }
 
     #[test]
     /// test that different type_next returns different addresses
     fn address_type_next() {
-        // @TODO is it possible to test that type_next changes the address in an isolated way?
-        // @see https://github.com/holochain/holochain-rust/issues/76
+        assert_ne!(
+            test_chain_header().address(),
+            ChainHeader::new(
+                &test_entry_type_a(),
+                &String::new(),
+                None,
+                &test_entry().address(),
+                &String::new(),
+                Some(test_chain_header().address())
+            ).address(),
+        );
     }
 
     /// Committing a LinkEntry to source chain should work
     #[test]
     fn can_round_trip_header_entry() {
-        let chain = test_chain();
-        let entry_type = test_entry_type();
-        let entry = test_entry();
-
-        let header = chain.create_next_chain_header(&entry_type, &entry);
-
-        let header_entry = header.to_entry().1;
-        let header_trip = ChainHeader::from_entry(&header_entry);
-
-        assert_eq!(header, header_trip);
+        assert_eq!(
+            test_chain_header(),
+            ChainHeader::from_entry(&test_chain_header().to_entry().1)
+        );
     }
 }
-*/
